third_party/WebKit/LayoutTests/http/tests/bluetooth/https/requestDevice/cross-origin-iframe.html

<!DOCTYPE html>
<script src="/js-test-resources/testharness.js"></script>
<script src="/js-test-resources/testharnessreport.js"></script>
<script src="/js-test-resources/testdriver.js"></script>
<script src="/js-test-resources/testdriver-vendor.js"></script>
<script src="/js-test-resources/bluetooth/bluetooth-helpers.js"></script>
<body>
<script>
'use strict';
const test_desc = 'Request device from a unique origin. ' +
  'Should reject with SecurityError.';

let iframe = document.createElement('iframe');

async_test(test =>
  setBluetoothFakeAdapter('HeartRateAdapter')
  // 1. Load the iframe.
  .then(() => new Promise(resolve => {
    iframe.src = 'https://localhost:8443/js-test-resources/bluetooth/requestDevice-in-iframe.html';
    document.body.appendChild(iframe);
    iframe.addEventListener('load', resolve);
  }))
  // 2. Request the device from the iframe.
  .then(() => {
    callWithTrustedClick(() => {
      iframe.contentWindow.postMessage('Go', '*');
    });

    window.onmessage = messageEvent => test.step(() => {
      assert_equals(messageEvent.data, 'SecurityError: requestDevice() ' +
          'called from cross-origin iframe.');
      test.done();
    });
}), test_desc);
</script>
</body>
